modularization expanded
