+++ src/outdoor/__init__.py
@@ -1,11 +1,11 @@
 __version__ = '0.0.0'
 
-from .outdoor_core.OUTDOOR import solve_OptimizationProblem
+from .outdoor_core.outdoor_main import solve_OptimizationProblem
 
-from .outdoor_core.Block_Results import displayBasicResults
-from .outdoor_core.Block_Results import displayHeatBalanceResults
-from .outdoor_core.Block_DataSaving import Save_CaseStudy, save_dict_to_file, load_dict_from_file
+from .outdoor_core.utils.output import displayBasicResults
+from .outdoor_core.utils.output import displayHeatBalanceResults
+from .outdoor_core.utils.saver import Save_CaseStudy, save_dict_to_file, load_dict_from_file
 
-from .excel_wrapper.ExcelWrapper import get_DataFromExcel
+from .excel_wrapper.main import get_DataFromExcel
 
-from .outdoor_core.Block_Superstructure import Superstructure
+from .outdoor_core.superstructure_sytem.superstructure import Superstructure
