+++ src/outdoor/excel_wrapper/Wrapp_Processes.py
@@ -12,9 +12,19 @@ Created on Thu Mar 26 10:56:00 2020
 """
 
 import pandas as pd
-from ..outdoor_core.Block_Process import *
-from .Wrapp_UnitData import *
-import outdoor.excel_wrapper.Wrapping_Functions  as WF
+from ..outdoor_core.unit_operations.library.pool import ProductPool
+from ..outdoor_core.unit_operations.library.source import Source
+from ..outdoor_core.unit_operations.library.stoich_reactor import StoichReactor
+from ..outdoor_core.unit_operations.library.yield_reactor import YieldReactor
+from ..outdoor_core.unit_operations.library.distributor import Distributor
+from ..outdoor_core.unit_operations.library.splitter import Splitter
+from ..outdoor_core.unit_operations.library.furnace import HeatGenerator
+from ..outdoor_core.unit_operations.library.turbine import ElectricityGenerator
+
+
+from ..outdoor_core.unit_operations.superclasses.physical_process import PhysicalProcess
+from .wrapp_unit_data import *
+from . import wrapping_functions as WF
