+++ src/outdoor/outdoor_core/model/optimization_model.py
