+++ src/outdoor/outdoor_core/unit_operations/superclasses/virtual_process.py
@@ -1,0 +1,19 @@
+import math 
+
+from .process import Process
+
+
+
+#------------------------------------------------------------------------------
+#------------------------------------------------------------------------------
+#--------------------------VIRTUAL PROCESSES-----------------------------------
+#------------------------------------------------------------------------------
+#------------------------------------------------------------------------------
+
+class VirtualProcess(Process):
+    
+    def __init__(self, Name, UnitNumber, Parent=None, *args, **kwargs):
+        
+        super().__init__(Name, UnitNumber, Parent)
+        
+
